Finish Welcome container and edited index.html
Finished the welcome to website container but needs a bit of tweaks and editing in terms
of resizing.
Edited the index.html file, added a new div for the introduction title and introduction text
in order to put them in flex. also added the class introduction title and changed it from h1 to dev,
height and width of the image